requirements.txt
docker-py
pyYAML
Mako